+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.34.0
 
+- Disallow `_` in account names.
 - Use DOMPurify.
 
 ## 0.33.0